Incorporates the fix for the bug in the (T) correction with project=3 or GNO=1
New MRCCSD calculations should use set_up_T_corr=T in the MRCC context
to run the extra needed targets. Add new tests for this case.

Also fix a small bug in python interface: the data of the classes were global
and not particular for each instance. This does not affect the present cases.