Fix bugs that affect on of the tests in gfortran
The test CCSD-R12fscr-3C.s was not running properly for gfortran.
The cause was an error in the memory freeing within
formula/topo_set_unique.f: because the real size of the array
contraction%svertex might not be contraction%nvtx (why?)
deallocate was trying to free non-existing elements of svtx.

Also, svmap(3) was out of bound. The test njoined.eq.3 must be in a
different statement, since fortran might not follow the order of the
tests when verifying the conditions.