Make a more versatile oldref for multistate MRCC
For a multistate mrcc with the oldref option, first the ME_C0_list.da is
looked for and used; if not found, uses the separeted
ME_C0_<i_state>_list.da instead.

A new rule is also added, SET_STATE: similar to ADV_STATE, but set the
state of ME-lists or operators to a specific state.