New features in MRPT: amplitudes equations and third order
Adds the possibility of having different kinds of amplitudes equations.
This is triggered by the new input keyword method.MRCCPT2.ampl_type
PT2 is the default

Also changes which components of the third order term (CC-like) is
calculated and printed.
Now prints the terms that are supposed to be zero if the T operator
contains all excitations (and thus CC and CI like would be the same)