Fix issue with AB_sym of density matrix
This commit changes the AB_sym of the density matrix (GAM0)
from +1 to msc (as is in the Fortran version)

The version with AB_sym=msc gives correc results: the formulas after
factoring out the formula for the density matrix matches the version
without factorisation. Also matches the fortran version.

The previous version, with AB_sym=+1 gives different results, which
does not converge or congerges poorly.

At this moment, is not yet clear why the AB_sym should be msc for the density.