Use different redirection syntax
The "&>" operator that was previously used is per se a valid Bash
operator, but on some systems it could be misinterpreted as a separate
"&" and ">" which would cause the test runner to be executed in the
background. This would let the tests fail with weird messages.

To avoid this problem, "2>&1 >" is used instead of "&>", which gets
parsed correctly.