Fix bug in split_contr4
The sign for moving around vertices of the split must be determined before
the contractions of the split are removed. Interestingly, this has so far
not induced any errors (test suite runs through with and without). I will
create dedicated tests soon.
Main changes are in split_contr4(), in addition I added an argument to some
routines from this call tree to allow local debugging (switch on/off in
the driver form_factor_out)