Some fixes in response to gfortran 9 and beyond troubles
Mainly, if statements that potentially access non-initialized pointes
were replaced by safer versions. They used to work as most compilers
generate code that not execute further statements if the first statement
already gives a definite result for a Boolean expression.

This version of the code runs without errors with gfortran 12.2 if -O0
is requested. For optimized code, unfortunately still errors occur.
To be continued ...